🏗️ Moves Exporter to Subpackage
We want to break this up into tinier pieces, and the first step is putting it somewhere that gives us room to spread out in.
